requirements.txt
psycopg2-binary==2.9.10
python-telegram-bot==21.9
PyYAML==6.0.2
